Removes locks and unwires the healthchecks thread.
This could probably be considered V2 already, this implementation is reworked
to remove completely locks and shared state.
Since I haven't found any satistfying nXn channel available in rust, I've used
the nice chrossbeam library, and built on top of it a dispatcher module, which
will deliver all the messages from the input queue, to all the output queues.
Service repository is now the central access point for the queries to the
service handler list, and it will also communicate any changes to the other
threads. It's basically CQRS, and for now it has responsabilities for both
sides.
I'm still not 100% sure that this approach can work, but haven't found any
significant drawback on paper. It's very efficient, and should make development
 of plugins and additional things fairly easy and efficient.
 For example, the reaper could build its own view of the service handlers,
 by storing a map <pid> : <sh>, thus avoiding searching for the sh / pid keys.
 This PR adds an initial implementation, which unfortunately unwires the
 healthcheck module, and it is also fairly incomplete (I really wanted to get
 it compiled, runnning and working - which it is as of now). But there are few
  things that need some more love. Until now though, I'm very satisfied with
  this new version, because it has simplified lots of things.