Updates the environment section
The old design was nice because it allowed user to defined additional
environment variables in a very clean way. The issue is that if people will run
horust from inside docker,I think they will mostly use the environment
section for re-exporting stuff to the process, defined in the dockerfile or
docker-compose file.
So this new design, greatly improve that use case, which was impossible in the
previous way of doing things.