Less work between fork/exec, fixes issue with bus
I've managed to reduce the work around fork / exec even more. This is to keep
things as safe as possible. Also, after many rounds of testing where I've
basically done while true ; cargo test, I've figured that there was an issue
with the exit condition for the bus, which was stuck on the blocking recv call.
I'm not sure why that happens, but with this new and less clever approach the
issue is gone and tests are all passing nicely.