Uses map for runtime's repo and some flow fixes
Now I'm trying to move the system towards a state
where the updates of the status are only emitted by the runtime component.
All the other components, listen to the status changes events and react
accordingly, but they communicate via events that will trigger next state
changes in the runtime, which in turn will emit ServiceStatus changes
accordingly.
So the healthchecker, will emit events "healthy/ unhealthy" (it's in the todo
list) and won't trigger any state change. If needed, the runtime will emit
appropriate state changes.