New healthcheker :tada:
This is the new version of the healthchecker. Not exactly what I initially
envisioned, but it's close enough. Every time a new service is started,
a worker thread is spawned which will take care of running the checks throught
the whole service life. In this way, every check can go at its own pace.
We still have one thread which is the manager of the workers, which it's fair
enough to have.
What's cool about this, now the repo is gone. It's communicating with the
outside via the Healthchecks, and that's it.
Less logic, less errors (hopefully).
I've also fixed a test, and refactored the shutting down case from the status
change handler.