Refactoring, tests and remove unused code from bus
Hopefully nobody is using it outside of horust.
It was cool to keep this code in there because in this way the bus was a
more complete implementation, but horust always used it in broadcast
mode, so there is no point in keeping this untested code in the repo.

I also did some minor refactorings and increased the test coverage.